% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{any_sparse_in_v}
\alias{any_sparse_in_v}
\title{A internal function to check whether a vector v contains any sparse responses.}
\usage{
any_sparse_in_v(v, min_f = 5)
}
\arguments{
\item{v}{A vector of questionnaire response data.}

\item{min_f}{The minimum frequency desired for response category data.
Any response categories that have a frequency count less than min_f will
be considered by sparse (i.e., have potentially too little data).}
}
\value{
TRUE or FALSE depending on whether sparse responses are present.
}
\description{
This function checks the vector v for any sparse responses.
It will return TRUE or FALSE. If v contains only NA, then NA will be
returned by the function.
}
\note{
This is intended as an internal function to
the collapse() function. Use with care for other purposes.
}
\examples{
\dontrun{

sparse_vector <- c(1, 1, 1, 1, 1, 2, 2, 3, 3, 3, 3, 3, 4, 4, 5, 5)
any_sparse_in_v(sparse_vector)

}
}
\seealso{
collapse()
}
