% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamie-internals.R
\name{col_pairs_to_matrix}
\alias{col_pairs_to_matrix}
\title{col_pairs_to_matrix}
\usage{
col_pairs_to_matrix(x, FUN, ...)
}
\arguments{
\item{x}{A dataframe.}

\item{FUN}{A function to be applied to each pair of columns in x.}

\item{...}{Additional arguments to be passed to FUN.}
}
\value{
A matrix[i, j] based on all pairs of columns in x.
}
\description{
This function populates a matrix[i, j] based on the columns of
a dataframe where i and j are both indices for 1 to the number of columns in
x.
}
\details{
The input must be a dataframe or an error will result.
}
\examples{
\dontrun{
col_pairs_to_matrix(a_dataframe, NMI))
}
}
