% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{i_freq_min}
\alias{i_freq_min}
\title{An internal function to return an index from among sparse response
categories. This function tells collapse() which response category
to start with when merging.}
\usage{
i_freq_min(tbl, return_which = c("min", "max"))
}
\arguments{
\item{tbl}{A table, usually one that was made from a vector of response
categories.}

\item{return_which}{In situations in which more than one response category
is the least frequent (i.e., with the lowest frequency count), should the
function return the minimum response (i.e., the leftmost category if sorted in
ascending order) or the maximum response (i.e., the rightmost category if
sorted in ascending order). The default is "min" for "minimum" or the
"rightmost" response option. This parameter is only relevant when more than
one response category is tied for being the least frequent.}
}
\value{
A single numeric index in tbl that tells collapse() which response
category to start with.
}
\description{
This function looks at a table, finds indices of the
least frequent response category (or categories), and returns one
as a starting point for collapse().
}
\note{
This is intended as an internal function to
the collapse() function. Use with care for other purposes.
}
\examples{
\dontrun{

sparse_vector <- c(1, 1, 1, 1, 1, 2, 2, 3, 3, 3, 3, 3, 4, 4, 5, 5)
i_freq_min(table(sparse_vector))
table(sparse_vector) |> i_freq_min()

}
}
\seealso{
collapse()
}
