% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{i_sparse}
\alias{i_sparse}
\title{An internal function that returns indices of sparse response categories.}
\usage{
i_sparse(tbl, min_f = 5)
}
\arguments{
\item{tbl}{A table, usually made from a vector of self-report items.}

\item{min_f}{The threshold for non-sparse frequencies. Any category less
than min_f will be flagged as sparse.}
}
\value{
Numerical indices indicating where in the table a sparse category
is found.
}
\description{
An internal function that returns the indices (i) of sparse
response categories, given a table (tbl).
}
\note{
This is intended as an internal function to
the collapse() function. Use with care for other purposes.
}
\examples{
\dontrun{

sparse_vector <- c(1, 1, 1, 1, 1, 2, 2, 3, 3, 3, 3, 3, 4, 4, 5, 5, 6, 6, 6)
i_sparse(table(sparse_vector))
i_sparse(table(sparse_vector, 3))

# By coercion
i_sparse(sparse_vector)

}
}
\seealso{
collapse()
}
