% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{merge_w_left}
\alias{merge_w_left}
\title{An internal function to that takes a response category found in a 
sorted table of v,  at index i in this table, and merges it with
the adjacent category to the left at i - 1.}
\usage{
merge_w_left(v, i)
}
\arguments{
\item{v}{A vector of response categories.}

\item{i}{An index of the to-be-collapsed category. i must range from 2 to
the maximum number of response categories. For example, for response 
categories 1, 2, 3, 4, and 5, "1" is not possible because there is no place
to the left. Thus, i can be 2, 3, 4, or 5.}
}
\value{
A vector of collapsed response data.
}
\description{
After sorting the response of v and putting them into a table,
this function looks at the category in said table at position i and merges
it with the category to the left at position i - 1.
}
\note{
This is intended as an internal function to
the collapse() function. Use with care for other purposes.
}
\examples{
\dontrun{

sparse_vector <- c(1, 1, 1, 1, 1, 2, 2, 3, 3, 3, 3, 3, 4, 4, 5, 5)
merge_w_left(sparse_vector, 2)

}
}
\seealso{
collapse()

merge_w_right()
}
