% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{merge_w_right}
\alias{merge_w_right}
\title{An internal function to that takes a response category found in a
sorted table of v,  at index i in this table, and merges it with
the adjacent category to the right at i + 1.}
\usage{
merge_w_right(v, i)
}
\arguments{
\item{v}{A vector of response categories.}

\item{i}{An index of the to-be-collapsed category. i must range from 1 to
the maximum number of response categories minus 1. For example, for response
categories 1, 2, 3, 4, and 5, "5" is not possible because there is no place
to the right. Thus, i can be 1, 2, 3, or 4.}
}
\value{
A vector of collapsed response data.
}
\description{
After sorting the response of v and putting them into a table,
this function looks at the category in said table at position i and merges
it with the category to the right at position i + 1.
}
\note{
This is intended as an internal function to
the collapse() function. Use with care for other purposes.
}
\examples{
\dontrun{

sparse_vector <- c(1, 1, 1, 1, 1, 2, 2, 3, 3, 3, 3, 3, 4, 4, 5, 5)
merge_w_right(sparse_vector, 2)

}
}
\seealso{
collapse()

merge_w_left()
}
