% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{n_resp_cats}
\alias{n_resp_cats}
\title{An internal function that returns the number of possible response categories.}
\usage{
n_resp_cats(v)
}
\arguments{
\item{v}{A vector, usually of self-report items, and usually numeric.}
}
\value{
The number of response categories in v. Note that NA is not counted
as a response category
}
\description{
An internal function that returns the number of possible
response categories in a vector v.
}
\note{
This is intended as an internal function to
the collapse() function. Use with care for other purposes.
}
\examples{
\dontrun{

sparse_vector <- c(1, 1, 1, 1, 1, 2, 2, 3, 3, 3, 3, 3, 4, 4, 5, 5, 6, 6, 6)
n_resp_cats(sparse_vector)

}
}
\seealso{
collapse()

n_sparse()
}
