% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{n_sparse}
\alias{n_sparse}
\title{An internal function that returns the number of sparse response categories.}
\usage{
n_sparse(v, min_f = 5)
}
\arguments{
\item{v}{A vector, usually of self-report items, and usually numeric.}

\item{min_f}{The threshold for non-sparse frequencies. Any category less
than min_f will be flagged as sparse.}
}
\value{
The number of sparse responses (i.e., < min_f) in v.
}
\description{
An internal function that returns the number of sparse
response categories in a vector v.
}
\note{
This is intended as an internal function to
the collapse() function. Use with care for other purposes.
}
\examples{
\dontrun{

sparse_vector <- c(1, 1, 1, 1, 1, 2, 2, 3, 3, 3, 3, 3, 4, 4, 5, 5, 6, 6, 6)
n_sparse(sparse_vector)

}
}
\seealso{
collapse()

n_resp_cats()
}
