% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmi_matrix.R
\name{nmi_matrix}
\alias{nmi_matrix}
\title{nmi_matrix}
\usage{
nmi_matrix(x, ...)
}
\arguments{
\item{x}{A dataframe (e.g., questionnaire items)}

\item{...}{Additional arguments to be passed to NMI as part of the
aricode package.}
}
\value{
A matrix[i, j] with NMI for each pair of columns.
}
\description{
This function populates a matrix with
normalised mutual information (NMI) for each pair of
columns in the dataframe x.
}
\details{
The input must be a dataframe or coercible to a dataframe.
}
\examples{
\dontrun{
nmi_matrix(RRS))
}
}
