% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamie-internals.R
\name{recode_items_in_df}
\alias{recode_items_in_df}
\title{recode_items_in_df}
\usage{
recode_items_in_df(items_df, original, recoded)
}
\arguments{
\item{items_df}{A dataframe containing items to be recoded.}

\item{original}{A vector containing the original coding of the variable
(e.g., 1, 2, 3, 4).}

\item{recoded}{A vector containing the recoding of the variable
(e.g., 4, 3, 2, 1). "recoded" and "original" must be the same length.}
}
\value{
A dataframe of recoded data are returned (usually questionnaire items).
}
\description{
This function is used to recode items within a dataframe.
(e.g., 1 = 4, 2 = 3, 3 = 2, 4 = 1).
}
\details{
The input must be a dataframe or an error will result.
}
\examples{
\dontrun{
score_items_in_df(a_dataframe[some_items])
}
}
