% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomp.R
\name{pomp}
\alias{pomp}
\title{polyPA}
\usage{
pomp(raw_scores, min_possible, max_possible)
}
\arguments{
\item{data}{Categorical input of questionnaire items}

\item{replications}{The number of replications in the simulation}

\item{prc}{The percentile for obtaining the reference eigenvalues}
}
\value{
Empirical and reference eigenvalues
}
\description{
Conducts a parallel analysis using the method of Lubbe (2019).
The code in this function is adapted from the key reference (see below).
This function relies on the psych package to calculate the polychoric correlations.
}
\section{Key Reference}{
 \url{https://doi.org/10.1037/met0000171}
}

\examples{
\dontrun{

pa_results <- polyPA(hl[1:10], 500, .99)

}
}
