% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamie-internals.R, R/score_surveys.R
\name{score_surveys}
\alias{score_surveys}
\title{score_surveys}
\usage{
score_surveys(items, min_num_items = ncol(items))

score_surveys(items, min_num_items = ncol(items))
}
\arguments{
\item{items}{A dataframe of numeric  questionnaire items.\cr
Note: If items is not a dataframe, the function will
stop and return an error message.}

\item{min_num_items}{The minimum number of items needed to score the
questionnaire. If not enough items are present, the score will be NA.
If some items are missing, but the number of non-missing items is
min_num_items or higher, then the score will be prorated.
By default, all items are required}
}
\value{
Scores for each case (i.e., row)

Scores for each row
}
\description{
Takes a dataframe of numeric items and calculates a score
}
\examples{
\dontrun{
# score_surveys(rrs, 10)
# score_surveys(rrs, 8)
}
# score_surveys(rrs, 10)
# score_surveys(rrs, 8)
}
